connection scenarios improved
