+++ .gitignore
@@ -114,4 +114,3 @@ entrypoint.sh
 Jenkinsfile
 example.json
 application.dev.yml
-application.test.yml
