+++ tests/__init__.py
@@ -1,0 +1,2 @@
+from tests.integrationtests.connection.testdata.connection_test_data import ConnectionTestData
+
