+++ tests/integrationtests/connection/TestConnectionDatabaseResource.py
@@ -4,102 +4,50 @@ from unittest import TestCase
 
 import pytest
 
-from infrastructor.IocManager import IocManager
 from infrastructor.data.DatabaseSessionManager import DatabaseSessionManager
 from infrastructor.data.Repository import Repository
 from models.dao.connection.Connection import Connection
 from models.dao.connection.ConnectionDatabase import ConnectionDatabase
+from tests.integrationtests.common.TestManager import TestManager
+from tests.integrationtests.connection.testdata.connection_test_data import ConnectionTestData
 
 
 class TestConnectionDatabaseResuource(TestCase):
     def __init__(self, methodName='TestConnectionDatabaseResuource'):
         super(TestConnectionDatabaseResuource, self).__init__(methodName)
-
-        from infrastructor.api.FlaskAppWrapper import FlaskAppWrapper
-        from infrastructor.scheduler.JobScheduler import JobScheduler
-        os.environ["PYTHON_ENVIRONMENT"] = 'test'
-        root_directory = os.path.abspath(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), os.pardir, os.pardir, os.pardir))
-        IocManager.configure_startup(root_directory=root_directory,
-                                     app_wrapper=FlaskAppWrapper,
-                                     job_scheduler=JobScheduler)
-        self.client = IocManager.app.test_client()
-
-    def print_error_detail(self, data):
-        print(data['message'] if 'message' in data else '')
-        print(data['traceback'] if 'traceback' in data else '')
-        print(data['message'] if 'message' in data else '')
+        self.test_manager = TestManager()
 
     def insert_connection(self, request):
         # insert connection
-        data = json.dumps(request)
-        response = self.client.post(
-            '/api/Connection/ConnectionDatabase',
-            data=data,
-            content_type='application/json',
-        )
-        response_data = json.loads(response.get_data(as_text=True))
-        print(response_data)
-        assert response.status_code == 200
-        self.print_error_detail(response_data)
+        response_data = self.test_manager.api_client.post('/api/Connection/ConnectionDatabase', request)
         return response_data
 
     def update_connection(self, request):
-        data = json.dumps(request)
-        response = self.client.put(
-            '/api/Connection/ConnectionDatabase',
-            data=data,
-            content_type='application/json',
-        )
-        response_data = json.loads(response.get_data(as_text=True))
-        print(response_data)
-        assert response.status_code == 200
-        self.print_error_detail(response_data)
+        # update connection
+        response_data = self.test_manager.api_client.put('/api/Connection/ConnectionDatabase', request)
         return response_data
 
     def test_database_connection(self):
-        name = 'NewDatabaseConnection'
-        test_insert_input = {
-            "Name": name,
-            "ConnectionTypeName": "Database",
-            "ConnectorTypeName": "POSTGRESQL",
-            "Host": "string",
-            "Port": 0,
-            "Sid": "string",
-            "DatabaseName": "string",
-            "User": "string",
-            "Password": "string"
-        }
-        test_update_input = {
-            "Name": name,
-            "ConnectorTypeName": "MSSQL",
-            "Host": "Test",
-            "Port": 1550,
-            "Sid": "test",
-            "DatabaseName": "test",
-            "User": "test",
-            "Password": "test"
-        }
         expected = True
-        expected_name = name
         try:
-            response_data = self.insert_connection(test_insert_input)
+            response_data = self.insert_connection(ConnectionTestData.test_insert_input)
             assert response_data['IsSuccess'] == expected
-            assert response_data['Result']['Name'] == expected_name
-            response_data = self.update_connection(test_update_input)
+            assert response_data['Result']['Name'] == ConnectionTestData.test_insert_input['Name']
+            response_data = self.update_connection(ConnectionTestData.test_update_input)
             assert response_data['IsSuccess'] == expected
-            assert response_data['Result']['Database'][0]['ConnectorType'][0]['Name'] == "MSSQL"
+            assert response_data['Result']['Database'][0]['ConnectorType'][0]['Name'] == \
+                   ConnectionTestData.test_update_input["ConnectorTypeName"]
         except Exception as ex:
             assert True == False
         finally:
             # clean integration test operations
-            database_session_manager: DatabaseSessionManager = IocManager.injector.get(DatabaseSessionManager)
+            database_session_manager: DatabaseSessionManager = self.test_manager.ioc_manager.injector.get(DatabaseSessionManager)
 
             connection_database_repository: Repository[ConnectionDatabase] = Repository[ConnectionDatabase](
                 database_session_manager)
             connection_repository: Repository[Connection] = Repository[Connection](
                 database_session_manager)
-            connection = connection_repository.first(Name=name, IsDeleted=0)
+            connection = connection_repository.first(Name=ConnectionTestData.test_insert_input['Name'], IsDeleted=0)
             connection_database = connection_database_repository.first(Connection=connection, IsDeleted=0)
             database_session_manager.session.delete(connection_database)
             database_session_manager.session.delete(connection)
