+++ tests/integrationtests/connection/TestConnectorTypeResuource.py
@@ -3,21 +3,14 @@ import os
 from unittest import TestCase
 
 from infrastructor.IocManager import IocManager
+from tests.integrationtests.common.TestManager import TestManager
 
 
 class TestConnectorTypeResuource(TestCase):
 
     def __init__(self, methodName='TestConnectorTypeResuource'):
         super(TestConnectorTypeResuource, self).__init__(methodName)
-
-        from infrastructor.api.FlaskAppWrapper import FlaskAppWrapper
-        from infrastructor.scheduler.JobScheduler import JobScheduler
-        os.environ["PYTHON_ENVIRONMENT"] = 'test'
-        root_directory = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), os.pardir, os.pardir, os.pardir))
-        IocManager.configure_startup(root_directory=root_directory,
-                                     app_wrapper=FlaskAppWrapper,
-                                     job_scheduler=JobScheduler)
-        self.client = IocManager.app.test_client()
+        self.test_manager = TestManager()
 
     def print_error_detail(self, data):
         print(data['message'] if 'message' in data else '')
@@ -25,14 +18,5 @@ class TestConnectorTypeResuource(TestCase):
         print(data['message'] if 'message' in data else '')
 
     def test_get_connector_type(self):
-
-        response = self.client.get(
-            '/api/Connection/ConnectorType',
-            content_type='application/json',
-        )
-
-        response_data = json.loads(response.get_data(as_text=True))
-        print(response_data)
-        assert response.status_code == 200
-        self.print_error_detail(response_data)
+        response_data = self.test_manager.api_client.get('/api/Connection/ConnectorType')
         assert response_data['IsSuccess'] == True
