+++ tests/integrationtests/connection/testdata/connection_test_data.py
@@ -1,0 +1,22 @@
+class ConnectionTestData:
+    test_insert_input = {
+        "Name": "TestConnection",
+        "ConnectionTypeName": "Database",
+        "ConnectorTypeName": "POSTGRESQL",
+        "Host": "string",
+        "Port": 0,
+        "Sid": "string",
+        "DatabaseName": "string",
+        "User": "string",
+        "Password": "string"
+    }
+    test_update_input = {
+        "Name": "TestConnection",
+        "ConnectorTypeName": "MSSQL",
+        "Host": "Test",
+        "Port": 1550,
+        "Sid": "test",
+        "DatabaseName": "test",
+        "User": "test",
+        "Password": "test"
+    }
